+++ design-system/lib/index.js
@@ -1,0 +1,4 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var DesignSystem_1 = require("./DesignSystem");
+exports.DesignSystem = DesignSystem_1.DesignSystem;