+++ src/layouts/Faq.vue
@@ -3,28 +3,27 @@
     <NavBar/>
     <div id="faq" class="shared">
       <section class="section-split">
+        <div id="design-language-explained" class="one-third question">
+          <h3>What's the difference between design tokens and a design language?</h3>
+        </div>
+        <div class="two-thirds answer">
+          <p>A design language is made out of design tokens—but it's more than "just a list" of design tokens.
+          Think of your body—it's a collection of cells, but more than "just a bunch" of cells.  The whole is greater than the sum of the parts.
+          <br><br>
+          When your design tokens are connected with structure, hierarchy, reusability, and modularity—built on the skeleton of a full programming language—
+          you can express more than just data.  You can express any visual identity, any look &amp; feel: any design language.</p>
+        </div>
         <div id="cross-platform-explained" class="one-third question">
           <h3>What platforms does Diez support?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>Diez has no intrinsic limitations in terms of target platforms or technologies. The current Diez beta includes compiler <router-link to="/glossary#targets">targets</router-link> for <strong>iOS SDKs</strong> written in Swift with Objective-C bridging, <strong>Android SDKs</strong> written in Kotlin, and <strong>Web SDKs</strong> written in JavaScript, TypeScript, CSS, and Sass. Most essentially, these are useful reference implementations of the Diez compiler API, which is arbitrarily extensible — our roadmap already includes additional configuration options for these compiler targets, and entirely new compiler targets.</p>
+          <p>Diez has no intrinsic limitations in terms of target platforms or technologies. Diez currently includes compiler <router-link to="/glossary#targets">targets</router-link> for <strong>iOS SDKs</strong> written in Swift with Objective-C bridging, <strong>Android SDKs</strong> written in Kotlin, and <strong>Web SDKs</strong> written in JavaScript, TypeScript, CSS, and Sass. Most essentially, these are useful reference implementations of the Diez compiler API, which is arbitrarily extensible — our roadmap already includes additional configuration options for these compiler targets, and entirely new compiler targets.</p>
         </div>
         <div class="one-third question">
           <h3>How is this different from React Native, Flutter, et al?</h3>
         </div>
         <div class="two-thirds answer">
           <p>Diez doesn't force your app to work within its paradigm or have opinions on how you design, develop, and structure your app's code. It simply provides you with SDKs for use in your existing codebases. Your generated Diez SDKs are libraries <strong>not frameworks</strong> — they are intended to complement, but not replace, your design workflow.</p>
-        </div>
-        <div id="extensibility-explained" class="one-third question">
-          <h3>How is Diez designed for extensibility?</h3>
-        </div>
-        <div class="two-thirds answer">
-          <p>The Diez toolkit consists of several layers, each equipped with flexible APIs:</p>
-          <ul>
-            <li>A <strong>command line interface</strong> which can be used to invoke “<span class="to-do-link-needed">commands</span>”. Anyone can provide additional commands in the command line interface, or add additional options to “<span class="to-do-link-needed">core</span>” commands such as <code class="inline">diez compile</code>.</li>
-            <li>An <strong>engine layer</strong>, provides an abstract definition of “<span class="to-do-link-needed">components</span>” which can be implemented to define cross-platform design systems.</li>
-            <li>A <strong>compiler layer</strong> that is capable of parsing user-defined and <router-link to="/glossary#prefabs">prefab</router-link> components and emitting native SDKs according to a flexible set of rules. With the Diez compiler API, you can define entirely new <router-link to="/glossary#targets">target</router-link> SDK languages and platforms, extend existing language/platform implementations, and, if needed, even provide custom <router-link to="/glossary#bindings">native bindings</router-link> for special components.</li>
-          </ul>
         </div>
         <div class="one-third question">
           <h3>How can I contribute to Diez?</h3>
@@ -36,13 +35,13 @@
           <h3>Do I have to use the components provided by Diez?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>No. You can use as many or as few of them as you want. You can design your own — Diez is designed for extensibility. Our current beta includes a number of <router-link to="/glossary#prefabs">prefab</router-link> components implementing some <router-link to="/glossary#tokens">design tokens</router-link> we've observed in design systems. We'll certainly be adding more, and you can too.</p>
+          <p>No. You can use as many or as few of them as you want. You can design your own — Diez is designed for extensibility. Diez currently includes a number of <router-link to="/glossary#prefabs">prefab</router-link> components implementing the most common <router-link to="/glossary#tokens">design tokens</router-link> we've observed in the wild. We'll certainly be adding more, and you can too.</p>
         </div>
         <div class="one-third question" id="where-it-lives">
           <h3>Where does the Diez source of truth live?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>This is up to you. We've found success with this pattern: The Diez project lives in a stand-alone GitHub repo and compiles SDKs into its consumer codebases.</p>
+          <p>This is up to you. We've found success with this pattern: The Diez project lives in a stand-alone GitHub repo and compiles SDKs into its consumer codebases.  We recommend GitHub's new [package manager](https://help.github.com/en/github/managing-packages-with-github-packages/about-github-packages) as an excellent way to distribute &amp; version your native packages between projects.</p>
         </div>
         <div class="one-third question" id="typescript">
           <h3>What is TypeScript and what does it have to do with Diez?</h3>
@@ -61,32 +60,32 @@
           <h3>Am I locked into Diez once I start using it for my team's design system?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>In production mode, you already are ejected from Diez. You own your design system, you own your SDK, and don't have to continue using Diez to update it if at any point you decide that it's not the right fit.</p>
+          <p>In production mode, you already are ejected from Diez.  Diez design languages simply boil down to data and code. You own your design language, you own your SDK, and don't have to continue using Diez to update it if at any point you decide that it's not the right fit.</p>
         </div>
         <div class="one-third question">
           <h3>How do I manage my UI components with Diez?</h3>
         </div>
         <div class="two-thirds answer">
-            <p>You don't. Diez doesn't currently handle any rendering, by design. Diez is a layer underneath your UI components — so if you maintain a React design component library, for example, you can hook it up to Diez, but it's up to you to maintain your UI component library. That said, this is something we're thinking about a lot. Have feedback any on this? <a href="https://spectrum.chat/diez/feedback">Let us know</a>!</p>
+            <p>You don't. Diez integrates seamlessly with existing UI component libraries and solutions (like Storybook), but is not intended to replace them.  Think of a complete design system as UI components plus a Design Language.  Diez is all about the latter.</p>
         </div>
         <div class="one-third question">
           <h3>Is Diez for designers as well as developers?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>Authoring Diez currently requires writing code.  It's straightforward to write, and at least as easy to write as React (probably easier!) — so if you consider yourself a low-coder, Diez is probably a great fit. We highly recommend authorizing Diez components using VS Code, which provides first-class support for TypeScript.</p>
+          <p>Diez is built on TypeScript, which is a programming language.  That said, we've carefully designed the API to be as JSON-like as possible for most authoring cases.  Generally, you'll want to be at least a "low-coder" to get a project set up, but anyone who's comfortable editing JSON or YAML should feel at home editing Diez.</p>
         </div>
         <div class="one-third question">
-          <h3>Does a team's design system source of truth live in design tools or in Diez?</h3>
+          <h3>How does Diez work with design files? (Sketch, Figma, etc.)</h3>
         </div>
         <div class="two-thirds answer">
-          <p>We envision treating design assets like (and as) code assets.  Think of design files as blobs of “alien source code,” which Diez compiles into a common language for cross-platform design systems. When we're able to build out more meaningful connections between design tools and Diez, the lines will get a bit blurrier, but for now, this is how we're approaching it.</p>
+          <p>Diez's Extractors pull design tokens out of design files.  What this means is you can collect design files & code into a holistic "source of truth" — where the common denominator is code.  Another way to put that is you can treat design files as "code artifacts," e.g. as packages you might install and import from a package manager like npm.</p>
         </div>
         <div id="hot-mode-explained" class="one-third question">
           <h3>How does hot updating work?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>Hot updating uses TypeScript's powerful ability to watch, typecheck, and hot-compile a codebase (in the case of Diez, a codebase containing Diez components) to observe and emit changes to your design system, then emit changes into your codebase. Under the hood, we are running a hot server based on the popular and powerful ability for web applications to implement “hot module reloading”, which basically means that we can react to and emit changes in real time, even if you're targeting a native codebase.</p>
-          <p>As a convenience for emitting changes to your design system into a native app while it's still running on your device or simulator, we are using webviews in <router-link to="/glossary#hot-mode">hot mode</router-link>, <em>only</em>. We plan to revise this approach in the future, but it's perfectly suitable for development and “melts away” in production builds.</p>
+          <p>Hot updating uses TypeScript's powerful ability to watch, typecheck, and hot-compile a codebase (in the case of Diez, a codebase containing Diez components) to observe and emit changes to your design language, then emit changes into your codebase. Under the hood, we are running a hot server based on the popular and powerful ability for web applications to implement “hot module reloading”, which basically means that we can react to and emit changes in real time, even if you're targeting a native codebase.</p>
+          <p>As a convenience for emitting changes to your design language into a native app while it's still running on your device or simulator, we are using webviews in <router-link to="/glossary#hot-mode">hot mode</router-link>, <em>only</em>. We plan to revise this approach in the future, but it's perfectly suitable for development and “melts away” in production builds.</p>
         </div>
         <div class="one-third question">
           <h3>Does hot updating (hot mode) affect App Store acceptability?</h3>
@@ -98,7 +97,7 @@
           <h3>Is hot updating necessary or required?</h3>
         </div>
         <div class="two-thirds answer">
-          <p><router-link to="/glossary#hot-mode">Hot mode</router-link> is a completely optional feature — you can use Diez-generated SDKs as a single source of truth for your design system without touching it at all.</p>
+          <p><router-link to="/glossary#hot-mode">Hot mode</router-link> is a completely optional feature.</p>
         </div>
         <div class="one-third question">
           <h3>Why would you want to use hot updating in development?</h3>
@@ -110,20 +109,7 @@
           <h3>Who's stewarding Diez?</h3>
         </div>
         <div class="two-thirds answer">
-          <p>Diez is a soon-to-be open source project designed and built by <a href="http://www.haikuforteams.com/">haikuforteams.com</a>.</p>
-        </div>
-        <div class="one-third question">
-          <h3>Is Diez ready for production usage, or is it still in beta?</h3>
-        </div>
-        <div class="two-thirds answer">
-          <p>Diez is in beta, but we are already having success using Diez to build our own apps. Diez is built with a high quality codebase, has excellent test coverage, and is always getting better.</p>
-        </div>
-        <div class="one-third question">
-          <h3>What's next for Diez?</h3>
-        </div>
-        <div class="two-thirds answer">
-          <p>We have an ambitious and exciting road ahead of us, and are very open to <a href="https://spectrum.chat/diez/feedback">feedback about which features you'd like to see</a> us work on next.</p>
-          <p>Our current roadmap will be on GitHub very soon.</p>
+          <p>Diez was created and is maintained by <a href="http://www.haikuforteams.com/">the Haiku Team</a>.</p>
         </div>
       </section>
     </div>