+++ src/layouts/Home.vue
@@ -10,8 +10,8 @@
             <HeroAnimation sizing="cover" loop="true"></HeroAnimation>
           </div>
           <div class="half align-top punch">
-            <h1>Connect any codebase to a shared design language</h1>
-            <h3>Create a shared repository for styles, assets, and usage guidelines—then compile to native code for iOS, Android, and Web.</h3>
+            <h1>The Design Language Framework</h1>
+            <h3>Diez radically reduces the cost of delivering a consistent visual identity across your company’s apps &amp; websites.</h3>
             <router-link class="button left" to="/getting-started">Get Started</router-link>
             <a class="button secondary left" :href="typeformUrl">Enterprise Edition</a>
           </div>
@@ -20,14 +20,17 @@
 
       <section id="what" class="section section-narrow">
         <h1 class="center heading1">What is Diez?</h1>
-        <h3 class="center">Diez is a free & open-source developer toolkit for creating codebase-ready design languages. Diez comprises three pieces:</h3>
+        <h3 class="center">Diez is a free &amp; open-source developer toolkit for expressing visual styles that can be shared across codebases, native platforms, and teams.
+          <br><br>
+          <strong>Diez comes in three parts:</strong>
+        </h3>
       </section>
 
       <section id="dlf" class="section-split">
         <div  class="half align-top">
           <h2>1. Design language framework</h2>
-          <p>Compose <a href="https://twitter.com/designtokens" target="_blank">design tokens</a> into any design language using Diez’s TypeScript framework.</p>
-          <p>From colors to typography, drop shadows to border radii, gradients to usage guidelines, logos to icons — if you can imagine it, you can express it in Diez.</p>
+          <p>Use TypeScript to compose <a href="/glossary/#tokens">design tokens</a> into <strong>scalable</strong> and <strong>extensible</strong> collections of styles, themes, and visual guidelines.  Your brand identity fits right here — as code.</p>
+          <p>From colors to typography, drop shadows to border radii, gradients to usage guidelines, logos to icons: if you can imagine it, you can express it in Diez.</p>
         </div>
         <div class="half">
           <img width="100%" src="@theme/assets/imgs/dlf.png" alt="diez compiler">
@@ -45,15 +48,14 @@
         <div class="half align-top">
           <h2>2. Cross-platform native compiler</h2>
           <p>Build any Diez project into pure native libraries for iOS, Android, or Web, ready for any new or existing codebase.</p>
-          <p>Version your design language as code packages with first-class support for CocoaPods, Carthage, Gradle, npm, and yarn.</p>
-          <p>Optimized for native-developer ergonomics — with autocomplete, inline comments, and static typing for iOS, Android, and Web, Diez makes it easy for everyone to rally around a common design language.</p>
+          <p>Optimized for native-developer ergonomics — with autocomplete, inline documentation, and package manager support for iOS, Android, and Web: Diez makes it easy for everyone to rally around a common design language.</p>
         </div>
       </section>
 
       <section id="extractors" class="section-split">
         <div class="half align-top">
           <h2>3. Design file extractors</h2>
-          <p>Diez Extractors compile design files into Diez-ready TypeScript, extracting colors, typography, image assets and more.</p>
+          <p>Diez Extractors compile design files into Diez-ready TypeScript — pulling colors, typography, image assets and more right out of your favorite design tools.</p>
           <p>Automate your design-to-development workflow by treating design files as versionable code artifacts.</p>
         </div>
         <div class="half mobile-reorder align-top">
@@ -66,8 +68,8 @@
       <section id="enterprise-grade">
         <div class="section section-narrow no-bottom">
           <h1 class="center">Built for Enterprise</h1>
-          <h3 class="center">Your Design System’s missing half</h3>
-          <p class="center">Diez naturally integrates with any UI component solution in the world, including the ones your team already uses.</p>
+          <h3 class="center">Your design system’s missing half</h3>
+          <p class="center">Diez naturally integrates with any UI component solution in the world to help you create a <strong>complete design system.</strong></p>
         </div>
         <div class="section full-img">
           <img src="@theme/assets/imgs/design-system.png" alt="design system" class="hide-on-mobile">
@@ -77,10 +79,10 @@
         <div class="section-split">
           <div class="half align-top">
             <h2>Scales like a dream</h2>
-            <p>Import, reuse & reference — great for composing multiple brands with shared assets & styles.</p>
+            <p>Import, reuse & reference: great for composing multiple brands with shared assets & styles.</p>
             <p>Strongly typed from end-to-end, Diez scales to thousands of files, hundreds of codebases, and dozens of teams without skipping a beat.</p>
             <h3>Infinitely customizable</h3>
-            <p>Diez is built around plugins — from the platforms it supports to the kinds of data you can express.  No matter the requirements of your design language or design system — whether you’re wrangling one brand or a hundred — you can customize Diez to fit your needs.</p>
+            <p>Diez is built around plugins: from the platforms it supports to the kinds of data you can express.  No matter the requirements of your design system — whether you’re wrangling one brand or a hundred — you can customize Diez to fit your needs.</p>
           </div>
           <div class="half">
             <img width="90%" src="@theme/assets/imgs/complexity.png" alt="complexity" class="mobile-img-spacer">
@@ -97,9 +99,9 @@
           </div>
         </div>
         <div class="one-third align-top">
-          <h2>DocsGen</h2>
+          <h2>DocsGen Add-on</h2>
           <h3>(Enterprise Edition only)</h3>
-          <p style="text-align: left;">From code comments and static analysis, Diez can automatically build customizable, ready-to-deploy documentation for your design system.</p>
+          <p style="text-align: left;">From any Diez project, Diez DocsGen builds up-to-the-moment, customizable, and self-hostable design language docs as static HTML pages. <br><br> It's never been easier to get everyone in your organization speaking the same design language.</p>
         </div>
       </section>
     </div>
@@ -107,7 +109,7 @@
     <div id="enterprise" class="shared">
       <section class="section">
         <h1>Community & Enterprise Editions</h1>
-        <p>Diez is open source, but has an Enterprise Edition with exclusive access to commercial add-ons, including DocsGen, Enterprise Training & Custom Integration options.</p>
+        <p>Diez is available in two editions:  the free and open source <strong>Diez Community Edition</strong> with everything you need to create your design language source of truth — and <strong>Diez Enterprise Edition</strong> with exclusive access to DocsGen, customizable commercial licensing, and other enterprise value-adds.</p>
         <div class="side-by-side">
           <div class="card">
             <h3>Community Edition</h3>
@@ -125,12 +127,12 @@
               <li>Design language framework</li>
               <li>Diez Compiler</li>
               <li>Design file extractors <span class="primary">(Including support for InVision DSM)</span></li>
-              <li class="primary">DocsGen</li>
+              <li class="primary">DocsGen Add-on</li>
               <li class="primary">Commercial license</li>
               <li class="primary">Support, Integration, & Training</li>
             </ul>
             <a class="button" :href="typeformUrl">
-              Contact Sales
+              Get in Touch
             </a>
             <div class="background"></div>
           </div>
@@ -368,6 +370,13 @@ export default {
 
   #enterprise-grade {
     margin-top: $spacing-xxxl-px;
+  }
+
+  #dlf {
+    margin-top: -70px;
+    @include phone {
+      margin-top: -20px;
+    }
   }
 
   #docsgen {